fix(sb/components/pages): move shuttleMapPage into pages/
